refact: rename protocol:deactivate to contract:deactivate
And rename `src/cli/protocol` to `src/cli/contract`.
