+++ src/cli/contract/activate/activate.js
