+++ src/cli/contract/activate/index.js
