+++ src/cli/contract/compile/compile.js
