+++ src/cli/contract/compile/index.js
