+++ src/cli/contract/index.js
