+++ src/cli/contract/list/index.js
