+++ src/cli/contract/list/list.js
