+++ src/cli/index.js
@@ -5,7 +5,7 @@ exports.blocklet = require('./blocklet');
 exports.chain = require('./chain');
 exports.misc = require('./misc');
 exports.node = require('./node');
-exports.protocol = require('./protocol');
+exports.protocol = require('./contract');
 exports.release = require('./release');
 exports.tools = require('./tools');
 exports.tx = require('./tx');
